Adding all source, header, and image files used
All header and source files needed to run the program are included as well as all the image files that we used to define sfml sprites. A cMakeLists file is also included which contains the script require to interact with the sfml library from Clion. SFML is required to run this code.